+++ Sources/VCoreMacrosImplementation/Macros/Freestanding (Expression)/Color Macro (Init with Hex UInt)/ColorMacro_InitWithHexUInt.swift
@@ -16,19 +16,27 @@ struct ColorMacro_InitWithHexUInt: ExpressionMacro {
         of node: some FreestandingMacroExpansionSyntax,
         in context: some MacroExpansionContext
     ) throws -> ExprSyntax {
-        // Parameters
-        let colorSpaceString: String = try colorSpaceParameter(node: node)
-        let hex: UInt = try hexParameter(node: node)
-        let opacity: CGFloat = try opacityParameter(node: node)
-
+        // Macro parameters
+        let colorSpaceString: String = try colorSpaceParameter(
+            node: node
+        )
+        
+        let hex: UInt = try hexParameter(
+            node: node
+        )
+        
+        let opacity: CGFloat = try opacityParameter(
+            node: node
+        )
+        
         // RGB values
         guard
             let rgbValues = hex._hexColorRGBValues()
         else {
-            throw ColorMacroError_InitWithHexUInt.invalidHexParameter
+            throw RawStringError("Invalid 'hex' parameter")
         }
 
-        // Result
+        // Macro expansion result
         return
             """
             Color(
@@ -59,7 +67,7 @@ struct ColorMacro_InitWithHexUInt: ExpressionMacro {
                 .baseName
                 .trimmedDescription
         else {
-            throw ColorMacroError_InitWithHexUInt.invalidColorSpaceParameter
+            throw RawStringError("Invalid 'colorSpace' parameter")
         }
 
         return value
@@ -73,7 +81,7 @@ struct ColorMacro_InitWithHexUInt: ExpressionMacro {
                 .arguments
                 .first(where: { $0.label?.trimmedDescription == "hex" })
         else {
-            throw ColorMacroError_InitWithHexUInt.invalidHexParameter
+            throw RawStringError("Invalid 'hex' parameter")
         }
 
         guard
@@ -85,7 +93,7 @@ struct ColorMacro_InitWithHexUInt: ExpressionMacro {
 
             let value: UInt = .init(valueString, radix: 16)
         else {
-            throw ColorMacroError_InitWithHexUInt.invalidHexParameter
+            throw RawStringError("Invalid 'hex' parameter")
         }
 
         return value
@@ -110,7 +118,7 @@ struct ColorMacro_InitWithHexUInt: ExpressionMacro {
 
             let value: CGFloat = Double(valueString).map({ CGFloat($0) })
         else {
-            throw ColorMacroError_InitWithHexUInt.invalidOpacityParameter
+            throw RawStringError("Invalid 'opacity' parameter")
         }
 
         return value